caching strategy
the original cache.go didn't get committed before it got deleted, so this is the best approximation from the explainer description we told the nerds about. we're essentially storing the request with a modtime header, then comparing the hash and from there comparing the modtime header so we know when to invalidate.